Fully implemented sampler

+++ src/sampler.rs
@@ -1,7 +1,7 @@
 use crate::configuration;
 
 pub trait Sampler{
-    fn generate_samples(&self, x: f64, y: f64) -> [[f64;2];(configuration::samples*configuration::samples) as usize];
+    fn generate_samples(&self, x: u32, y: u32) -> [[f64;2];(configuration::samples*configuration::samples) as usize];
     fn evaluate(&mut self, t: f64);
 }
 
@@ -11,7 +11,7 @@ pub struct JitterSampler{
 const SAMPLE_INV: f64 = 1.0 / (configuration::samples as f64);
 
 impl Sampler for JitterSampler{
-    fn generate_samples(&self, x: f64, y: f64) -> [[f64;2];(configuration::samples*configuration::samples) as usize]{
+    fn generate_samples(&self, x: u32, y: u32) -> [[f64;2];(configuration::samples*configuration::samples) as usize]{
         let mut elements = [[0.0, 0.0];(configuration::samples*configuration::samples) as usize];
         
         for sx in 0..configuration::samples{
